Add dependency on responders gem
In Rails 4.2, responders were moved out of core and into a gem, so this
dependency is required. It should be backwards compatible for all Rails
3.2+ versions.

As part of this change, the Responder class--and the workaround in the
README for Rails 4.2--needed to be changed to call #body on the Response
object.